requirements.txt
absl-py==0.1.13
astor==0.6.2
backports.weakref==1.0.post1
bleach==1.5.0
boto3==1.7.4
botocore==1.10.4
certifi==2018.1.18
chardet==3.0.4
click==6.7
dicom==0.9.9.post1
docutils==0.14
enum34==1.1.6
Flask==0.12.2
funcsigs==1.0.2
futures==3.2.0
gast==0.2.0
gevent==1.2.2
greenlet==0.4.13
grpcio==1.11.0
gunicorn==19.7.1
h5py==2.7.1
html5lib==0.9999999
idna==2.6
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
Keras==2.1.5
Markdown==2.6.11
MarkupSafe==1.0
mock==2.0.0
mritopng==2.1
numpy==1.14.2
opencv-python==3.4.0.12
pbr==4.0.2
Pillow==5.1.0
protobuf==3.5.2.post1
pydicom==1.0.2
pypng==0.0.18
python-dateutil==2.6.1
PyYAML==3.12
requests==2.18.4
s3transfer==0.1.13
scipy==1.0.1
six==1.11.0
tensorboard==1.7.0
tensorflow==1.7.0
termcolor==1.1.0
urllib3==1.22
Werkzeug==0.14.1
